日常保存
Damage

+++ Fight Sys Demo/Assets/Scripts/Damage/Damage.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class Damage
 {
     public DamageType damageType = DamageType.PHYSICS;
-    public int value = 1;
+    public float value = 1;
 
-    public Damage(DamageType damage_type, int value)
+    public Damage(DamageType damage_type, float value)
     {
         this.damageType = damage_type;
         this.value = value;
